mp3_frame_tags: Covert to decode group and split to mp3_frame_{xing,vbri} decoders
