+++ README.md
@@ -106,7 +106,8 @@ macho_fat,
 [matroska](doc/formats.md#matroska),
 [mp3](doc/formats.md#mp3),
 mp3_frame,
-mp3_frame_tags,
+mp3_frame_vbri,
+mp3_frame_xing,
 [mp4](doc/formats.md#mp4),
 mpeg_asc,
 mpeg_es,
