+++ format/all/all.fqtest
@@ -109,7 +109,8 @@ markdown             Markdown
 matroska             Matroska file
 mp3                  MP3 file
 mp3_frame            MPEG audio layer 3 frame
-mp3_frame_tags       MP3 frame info/xing tags
+mp3_frame_vbri       MP3 frame Fraunhofer encoder variable bitrate tag
+mp3_frame_xing       MP3 frame Xing/Info tag
 mp4                  ISOBMFF, QuickTime and similar
 mpeg_asc             MPEG-4 Audio Specific Config
 mpeg_es              MPEG Elementary Stream
