+++ pkg/interp/testdata/slurp.fqtest
@@ -94,7 +94,7 @@ mp3> $f[]
 0x30|   00 00 00 00 00 00 00 00 00 00 00 00 00 00 00| ...............|  side_info{}:
 0x40|00 00                                          |..              |
     |00 01 02 03 04 05 06 07 08 09 0a 0b 0c 0d 0e 0f|0123456789abcdef|
-0x40|      49 6e 66 6f 00 00 00 0f 00 00 00 02 00 00|  Info..........|  tag{}: (mp3_frame_tags)
+0x40|      49 6e 66 6f 00 00 00 0f 00 00 00 02 00 00|  Info..........|  tag{}: (mp3_frame_xing)
 0x50|02 57 00 a6 a6 a6 a6 a6 a6 a6 a6 a6 a6 a6 a6 a6|.W..............|
 *   |until 0xdd.7 (156)                             |                |
 0xd0|                                          00 00|              ..|  audio_data: raw bits
